requirements.txt
Flask==3.0.1
Flask_SocketIO==5.3.6
mediapipe==0.10.9
numpy==1.24.3
opencv_contrib_python==4.9.0.80
opencv_python==4.8.1.78
Pillow==10.2.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
